requirements.txt
babel==2.18.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.2
pytz==2026.1.post1
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6
WTForms==3.2.1
